Move prop types into component file
