--- packages/react-day-picker/src/components/Button/ButtonProps.tsx
@@ -1,4 +1,0 @@
-import * as React from 'react';
-
-/** The props for the [[Button]] component. */
-export type ButtonProps = React.HTMLProps<HTMLButtonElement>;
