+++ packages/react-day-picker/src/components/Caption/Caption.tsx
@@ -10,7 +10,11 @@ import { YearsDropdown } from '../YearsDropdown';
 import { useDayPicker } from '../../contexts/DayPicker';
 import { useNavigation } from '../../contexts/Navigation';
 
-import { CaptionProps } from './CaptionProps';
+/** Represent the props of the [[Caption]] component. */
+export interface CaptionProps {
+  /** The month where the caption is displayed. */
+  displayMonth: Date;
+}
 
 /**
  * Render the caption of a month, which includes title and navigation buttons.
