--- packages/react-day-picker/src/components/Caption/CaptionProps.ts
@@ -1,5 +1,0 @@
-/** Represent the props of the [[Caption]] component. */
-export interface CaptionProps {
-  /** The month where the caption is displayed. */
-  displayMonth: Date;
-}
