+++ packages/react-day-picker/src/components/Caption/index.ts
@@ -1,2 +1,1 @@
 export * from './Caption';
-export * from './CaptionProps';
