--- packages/react-day-picker/src/components/CaptionLabel/CaptionLabelProps.ts
@@ -1,5 +1,0 @@
-/** The props for the [[CaptionLabel]] component. */
-export interface CaptionLabelProps {
-  /** The month where the caption is displayed. */
-  displayMonth: Date;
-}
