+++ packages/react-day-picker/src/components/CaptionLabel/index.ts
@@ -1,2 +1,1 @@
 export * from './CaptionLabel';
-export * from './CaptionLabelProps';
