+++ packages/react-day-picker/src/components/Day/Day.tsx
@@ -2,8 +2,15 @@ import * as React from 'react';
 
 import { Button } from '../Button';
 
-import { DayProps } from './DayProps';
 import { useDay } from './hooks/useDay';
+
+/** Represent the props used by the [[Day]] component. */
+export interface DayProps {
+  /** The month where the date is displayed. */
+  displayMonth: Date;
+  /** The date to render. */
+  date: Date;
+}
 
 /**
  * The content of a day cell – as a button or span element according to its
