--- packages/react-day-picker/src/components/Day/DayProps.ts
@@ -1,7 +1,0 @@
-/** Represent the props used by the [[Day]] component. */
-export interface DayProps {
-  /** The month where the date is displayed. */
-  displayMonth: Date;
-  /** The date to render. */
-  date: Date;
-}
