+++ packages/react-day-picker/src/components/Day/index.ts
@@ -1,3 +1,2 @@
 export * from './Day';
-export * from './DayProps';
 export * from './hooks';
