+++ packages/react-day-picker/src/components/DayContent/DayContent.tsx
@@ -2,7 +2,17 @@ import * as React from 'react';
 
 import { useDayPicker } from '../../contexts/DayPicker';
 
-import { DayContentProps } from './DayContentProps';
+import { ModifierStatus } from '../../types';
+
+/** Represent the props for the [[DayContent]] component. */
+export interface DayContentProps {
+  /** The date representing the day. */
+  date: Date;
+  /** The month where the day is displayed. */
+  displayMonth: Date;
+  /** The modifier status for the date. */
+  modifiers: ModifierStatus;
+}
 
 /**
  * Render the content of the day cell.
