+++ packages/react-day-picker/src/components/DayContent/index.ts
@@ -1,2 +1,1 @@
 export * from './DayContent';
-export * from './DayContentProps';
