+++ packages/react-day-picker/src/components/Dropdown/index.ts
@@ -1,2 +1,1 @@
 export * from './Dropdown';
-export * from './DropdownProps';
