+++ packages/react-day-picker/src/components/MonthsDropdown/index.ts
@@ -1,2 +1,1 @@
 export * from './MonthsDropdown';
-export * from './MonthsDropdownProps';
