+++ packages/react-day-picker/src/components/Navigation/Navigation.tsx
@@ -3,7 +3,23 @@ import * as React from 'react';
 import { Button } from '../Button';
 import { useDayPicker } from '../../contexts/DayPicker';
 
-import { NavigationProps } from './NavigationProps';
+/** The props for the [[Navigation]] component. */
+export interface NavigationProps {
+  /** The month where the caption is displayed. */
+  displayMonth: Date;
+  /** The previous month. */
+  previousMonth?: Date;
+  /** The next month. */
+  nextMonth?: Date;
+  /** Hide the previous button. */
+  hidePrevious: boolean;
+  /** Hide the next button. */
+  hideNext: boolean;
+  /** Event handler when the next button is clicked. */
+  onNextClick: React.MouseEventHandler<HTMLButtonElement>;
+  /** Event handler when the previous button is clicked. */
+  onPreviousClick: React.MouseEventHandler<HTMLButtonElement>;
+}
 
 /** A component rendering the navigation buttons or the drop-downs. */
 export function Navigation(props: NavigationProps): JSX.Element {
