+++ packages/react-day-picker/src/components/Navigation/index.ts
@@ -1,2 +1,1 @@
 export * from './Navigation';
-export * from './NavigationProps';
