+++ packages/react-day-picker/src/components/Row/Row.tsx
@@ -4,7 +4,17 @@ import { getUnixTime } from 'date-fns';
 
 import { useDayPicker } from '../../contexts/DayPicker';
 
-import { RowProps } from './RowProps';
+/**
+ * The props for the [[Row]] component.
+ */
+export interface RowProps {
+  /** The month where the row is displayed. */
+  displayMonth: Date;
+  /** The number of the week to render. */
+  weekNumber: number;
+  /** The days contained in the week. */
+  dates: Date[];
+}
 
 /**
  * Render a row in the calendar, with the days and optionally the week number.
