--- packages/react-day-picker/src/components/Row/RowProps.ts
@@ -1,11 +1,0 @@
-/**
- * The props for the [[Row]] component.
- */
-export interface RowProps {
-  /** The month where the row is displayed. */
-  displayMonth: Date;
-  /** The number of the week to render. */
-  weekNumber: number;
-  /** The days contained in the week. */
-  dates: Date[];
-}
