+++ packages/react-day-picker/src/components/Row/index.ts
@@ -1,2 +1,1 @@
 export * from './Row';
-export * from './RowProps';
